+++ lib/que/scheduler.rb
@@ -1,6 +1,7 @@
 require "que/scheduler/version"
 require "que/scheduler/version_support"
 require "que/scheduler/configuration"
+require "que/scheduler/time_zone"
 require "que/scheduler/scheduler_job"
 require "que/scheduler/db"
 require "que/scheduler/audit"