+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Add time_zone configuration option to simplify running outside of Rails
+
 ## 4.0.2 (2020-11-29)
 
 - Add method to reenqueue scheduler in squashed migrations [#226](https://github.com/hlascelles/que-scheduler/pull/226)